# openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: DomiHive Unified Backend API
  version: 1.0.0
  description: >
    Baseline v1 OpenAPI contract for DomiHive modular monolith backend,
    aligned to execution plan dated April 25, 2026.
servers:
  - url: http://localhost:4000
    description: Local backend
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Register account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
              required: [phone, password]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkEnvelope' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                password: { type: string }
              required: [phone, password]
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkEnvelope' }
...
  /api/v1/tenant/listings:
    get:
      tags: [Tenant]
      summary: List tenant browse listings
      parameters:
        - in: query
          name: state
          schema: { type: string }
        - in: query
          name: area
          schema: { type: string }
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: propertyType
          schema: { type: string }
      responses:
        '200':
          description: Listing feed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Listing' }
                          syncedAt:
                            type: string
                            format: date-time
  /api/v1/tenant/listings/{listingId}:
    get:
      tags: [Tenant]
      summary: Get listing details
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing item
  /api/v1/tenant/inspections:
    post:
      tags: [Tenant]
      summary: Book inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                listingId: { type: string }
                inspectionDate: { type: string, format: date }
                inspectionTime: { type: string }
                attendees: { type: integer, minimum: 1 }
              required: [listingId, inspectionDate, inspectionTime]
      responses:
        '201':
          description: Inspection booked
...
  /api/v1/tenant/inspections/{inspectionId}/reschedule:
    patch:
      tags: [Tenant]
      summary: Reschedule inspection
      parameters:
        - $ref: '#/components/parameters/InspectionId'
      responses:
        '200':
          description: Rescheduled
  /api/v1/tenant/applications:
    post:
      tags: [Tenant]
      summary: Create rental application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                listingId: { type: string }
              required: [listingId]
      responses:
        '201':
          description: Application created
    get:
      tags: [Tenant]
      summary: Get my applications
      responses:
        '200':
          description: Applications list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/RentalApplication' }
  /api/v1/tenant/applications/{applicationId}:
    get:
      tags: [Tenant]
      summary: Get application by id
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Application item
    patch:
      tags: [Tenant]
      summary: Update application
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Updated
  /api/v1/tenant/applications/{applicationId}/documents:
    post:
      tags: [Tenant]
      summary: Upload application document metadata
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '201':
          description: Uploaded
  /api/v1/tenant/applications/{applicationId}/payments:
    post:
      tags: [Tenant]
      summary: Submit application payment
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '201':
          description: Payment captured
...
  /api/v1/tenant/properties/{propertyId}/move-in:
    patch:
      tags: [Tenant]
      summary: Complete move-in checklist
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Move-in confirmed
...
  /api/v1/tenant/messages/threads/{threadId}/messages:
    post:
      tags: [Tenant]
      summary: Send message in thread
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      responses:
        '201':
          description: Message created
...
  /api/v1/admin/properties/{propertyId}:
    get:
      tags: [Admin]
      summary: Get property
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Property item
    patch:
      tags: [Admin]
      summary: Update property
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Property updated
  /api/v1/admin/properties/{propertyId}/units:
    post:
      tags: [Admin]
      summary: Create unit
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '201':
          description: Unit created
  /api/v1/admin/units/{unitId}:
    patch:
      tags: [Admin]
      summary: Update unit
      parameters:
        - $ref: '#/components/parameters/UnitId'
      responses:
        '200':
          description: Unit updated
...
  /api/v1/admin/applications/{applicationId}/status:
    patch:
      tags: [Admin]
      summary: Update application status
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/ApplicationStatus'
              required: [status]
      responses:
        '200':
          description: Status updated
...
  /api/v1/admin/inspections/{inspectionId}/status:
    patch:
      tags: [Admin]
      summary: Update inspection status
      parameters:
        - $ref: '#/components/parameters/InspectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/InspectionStatus'
              required: [status]
      responses:
        '200':
          description: Status updated
...
  /api/v1/admin/maintenance/{ticketId}:
    patch:
      tags: [Admin]
      summary: Update maintenance ticket
      parameters:
        - in: path
          name: ticketId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Ticket updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ListingId:
      in: path
      name: listingId
      required: true
      schema: { type: string }
    ApplicationId:
      in: path
      name: applicationId
      required: true
      schema: { type: string }
    InspectionId:
      in: path
      name: inspectionId
      required: true
      schema: { type: string }
    PropertyId:
      in: path
      name: propertyId
      required: true
      schema: { type: string }
    UnitId:
      in: path
      name: unitId
      required: true
      schema: { type: string }
...
    ThreadId:
      in: path
      name: threadId
      required: true
      schema: { type: string }
  schemas:
    OkEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
      required: [ok]
...
    ApplicationStatus:
      type: string
      enum:
        - INSPECTION_SCHEDULED
        - INSPECTION_VERIFIED
        - APPLICATION_STARTED
        - APPLICATION_SUBMITTED
        - UNDER_REVIEW
        - APPROVED
        - REJECTED
        - CANCELLED
    InspectionStatus:
      type: string
      enum:
        - PENDING_CONFIRMATION
        - SCHEDULED
        - NO_SHOW
        - INSPECTION_COMPLETED
...
    Listing:
      type: object
      properties:
        id: { type: string }
        listingId: { type: string }
        propertyId: { type: string }
        unitId: { type: string }
        unitCode: { type: string }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        cautionFee: { type: number }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        size: { type: string }
        state: { type: string }
        area: { type: string }
        locationName: { type: string }
        location: { type: string }
        address: { type: string }
        postalCode: { type: string }
        propertyType: { type: string }
        propertyTypeLabel: { type: string }
        buildingType: { type: string }
        buildingTypeLabel: { type: string }
        managementType: { type: string }
        isEstate: { type: boolean }
        propertyAge: { type: string }
        petsAllowed: { type: boolean }
        furnishing: { type: string }
        amenityIds:
          type: array
          items: { type: string }
        images:
          type: array
          items: { type: string }
        image: { type: string }
        tenantStatus: { type: string }
        canBook: { type: boolean }
        forRent: { type: boolean }
    RentalApplication:
      type: object
      properties:
        id: { type: string }
        listingId: { type: string }
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        inspectionStatus:
          $ref: '#/components/schemas/InspectionStatus'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }